Student - XML read, write from xml file

+++ WPFStudentSE1711/WPFStudent/MainWindow.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,6 +13,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using System.Xml.Serialization;
 using WPFStudent.Models;
 namespace WPFStudent
 {
@@ -87,22 +89,23 @@ namespace WPFStudent
 
         private void btnAdd_Click(object sender, RoutedEventArgs e)
         {
-            try
-            {
-                var student = GetStudentObject();
-                if(student != null)
-                {
-                    student.Id = 0;
-                    _context.Students.Add(student);
-                    _context.SaveChanges();
-                    Load_Data();
-                    MessageBox.Show("Insert student success " + student.Id +" "+ student.Name, "Add Employee");
-                }
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show("Insert student failed", "Add Student");
-            }
+            //try
+            //{
+            //    var student = GetStudentObject();
+            //    if(student != null)
+            //    {
+            //        student.Id = 0;
+            //        _context.Students.Add(student);
+            //        _context.SaveChanges();
+            //        Load_Data();
+            //        MessageBox.Show("Insert student success " + student.Id +" "+ student.Name, "Add Employee");
+            //    }
+            //}
+            //catch (Exception ex)
+            //{
+            //    MessageBox.Show("Insert student failed", "Add Student");
+            //}
+            btnAddClickV2();
         }
 
         private void btnUpdate_Click(object sender, RoutedEventArgs e)
@@ -140,13 +143,18 @@ namespace WPFStudent
                 var student = GetStudentObject();
                 if (student != null)
                 {
+                    
                     var oldStudent = _context.Students.FirstOrDefault(x => x.Id == student.Id);
                     if (oldStudent != null)
                     {
-                        _context.Students.Remove(oldStudent);
-                        _context.SaveChanges();
-                        Load_Data();
-                        MessageBox.Show("Delete student success " + student.Id + " " + student.Name, "Delete Student");
+                        var result = MessageBox.Show("Are you sure to delete " + oldStudent.Name, "Question", MessageBoxButton.YesNoCancel, MessageBoxImage.Question);
+                        if(result == MessageBoxResult.Yes)
+                        {
+                            _context.Students.Remove(oldStudent);
+                            _context.SaveChanges();
+                            Load_Data();
+                            MessageBox.Show("Delete student success " + student.Id + " " + student.Name, "Delete Student");
+                        }
                     }
                 }
             }
@@ -163,6 +171,83 @@ namespace WPFStudent
 
         }
 
+        //----------------------------------------
+        public Student GetStudentV2()
+        {
 
+            Student tmp = new Student()
+            {
+                Id = string.IsNullOrEmpty(txtId.Text) ? 0 : int.Parse(txtId.Text),
+                Name = txtName.Text,
+                Dob = txtDateofBirth.SelectedDate,
+                Gender = rdoMale.IsChecked == true ? true : false,
+                Major = Int32.Parse(cboNganh.SelectedValue.ToString())
+            };
+            return tmp;
+        }
+        public void btnAddClickV2()
+        {
+            try
+            {
+                Student tmp = GetStudentV2();
+                if (tmp != null)
+                {
+                    var p = _context.Students.FirstOrDefault(x => x.Id == tmp.Id);
+                    if (p == null)
+                    {
+                        _context.Students.Add(tmp);
+                        _context.SaveChanges();
+                        Load_Data();
+                    }
+                }
+
+            }
+            catch (Exception e)
+            {
+                MessageBox.Show(e.Message);
+            }
+        }
+
+        private void btnXML_Click(object sender, RoutedEventArgs e)
+        {
+            XmlSerializer xs = new XmlSerializer(typeof(XMLStudent));
+            string x = File.ReadAllText("XMLStudent.xml");
+            using Stream s2 = File.OpenRead("XMLStudent.xml");//đọc lại theo dạng stream 
+            var p2 = (XMLStudent)xs.Deserialize(s2);//giải mã trở lại thành đối tượng
+            DateTime dob;
+            if(p2!= null)
+            {
+                dob = System.DateTime.Parse(p2.Dob + " 12:00:00AM");
+                Student s = new Student()
+                {
+                    Name = p2.Name,
+                    Dob = dob,
+                    Gender = p2.Gender,
+                    Major = p2.Major,
+                };
+                _context.Students.Add(s);
+                _context.SaveChanges() ;
+                Load_Data() ;
+            }
+
+        }
     }
+    [XmlRoot("XMLStudent")]
+    public class XMLStudent
+    {
+        [XmlElement("Id")]
+        public string Id { get; set; }
+
+        [XmlElement("Name")]
+        public string Name { get; set; }
+
+        [XmlElement("Dob")]
+        public string Dob { get; set; }
+
+        [XmlElement("Gender")]
+        public bool Gender { get; set; }
+        
+        [XmlElement("Major")]
+        public int Major { get; set; }
+    }//end class
 }
